feat: add middleware on all route

+++ resources/views/class.blade.php
@@ -16,8 +16,10 @@
 
         <div class="d-flex justify-content-end my-3">
             <td>
-                <a href="class-add" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Tambah data</a>
-                <a href="class-deleted" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Show deleted class</a>
+                @if (Auth::user()->role_id == 1)
+                    <a href="class-add" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Tambah data</a>
+                    <a href="class-deleted" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Show deleted class</a>
+                @endif
             </td>
         </div>
 
@@ -54,8 +56,10 @@
                     </td> --}}
                     <td>
                         <a href="class/{{$clas->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Detail</a>
-                        <a href="class-edit/{{$clas->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Edit</a>
-                        <a href="class-delete/{{$clas->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Delete</a>
+                        @if (Auth::user()->role_id == 1)
+                            <a href="class-edit/{{$clas->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Edit</a>
+                            <a href="class-delete/{{$clas->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Delete</a>
+                        @endif
                     </td>
                 </tr>
