+++ resources/views/teacher.blade.php
@@ -17,8 +17,10 @@
 
         <div class="d-flex justify-content-end my-3">
             <td>
-                <a href="teacher-add" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Tambah data</a>
-                <a href="teacher-deleted" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Show data deleted</a>
+                @if (Auth::user()->role_id == 1)
+                    <a href="teacher-add" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Tambah data</a>
+                    <a href="teacher-deleted" class="btn btn-primary mx-2" tabindex="-1" role="button" aria-disabled="true">Show data deleted</a>
+                @endif
             </td>
         </div>
 
@@ -33,7 +35,9 @@
             <tr>
                 <th>No</th>
                 <th>Nama</th>
-                <th>Action</th>
+                @if (Auth::user()->role_id == 1)
+                    <th>Action</th>
+                @endif
             </tr>
 
             @foreach ($teachers as $teacher)
@@ -41,9 +45,11 @@
                     <td>{{ $loop->iteration }}</td>
                     <td>{{ $teacher->name }}</td>
                     <td>
-                        <a href="teacher/{{$teacher->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Detail</a>
-                        <a href="teacher-edit/{{$teacher->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Edit</a>
-                        <a href="teacher-delete/{{$teacher->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">delete</a>
+                        @if (Auth::user()->role_id == 1)
+                            <a href="teacher/{{$teacher->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Detail</a>
+                            <a href="teacher-edit/{{$teacher->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">Edit</a>
+                            <a href="teacher-delete/{{$teacher->id}}" class="btn btn-secondary mb-2" tabindex="-1" role="button" aria-disabled="true">delete</a>
+                        @endif
                     </td>
                 </tr>
             @endforeach
